my first commit
1. Using ErrorInfo class to handle error code and meassage .
When  caching  exceptions in the method of "downad()" , I get the
Meassges of the exception  and put it in a errorInfo ,then create a
DownloadThrowable that extends Throwable and contains the errorInfo.
2. In the class of DownloadManager ，I complete methods in the interface
of DownloadTaskListenner.
In the method of errorDownload ,I have  handled the queue of loadingtask
,the  cache of ConfigUitls and send broadcast .
When file is existing  i think it is not exception  it is  normal
process.